Reimplement message queues.
This time tested with gcc, so it is certain to compile with it.
Also, use a simpler model, making debugging easier.

Dropped size limiter (petrinets have mechanisms to enforce this in the model,
which is easier to prove).

Dropped reference counters: without reference counters, dedicated classes
can be created modeling a whole petrinet at once.