Implement tp_aid_service.
tp_aid_service is a simple service that provides do_work/has_work and allows
setting a wakeup-callback.  It's ideally suited for threads that sometimes want
to aid a client of a threadpool, but cannot dedicate all their time
(for instance, to be mixed in with the main event loop of a program).