Implement threadpool.
--

Provide pre-compiled implementations of threadpool_attach.
Reduces size of generated code in users of this library.

--

Clear threadpool service on service destruction.

This keeps the service available while the threads complete
(because they copy the reference pointer)
but blocks new loop-invocations on the threadpool
from performing work for the client.

Since the lifetime of the worker threads is limited to the lifetime of
the threadpool, the service is available while the worker threads use
the threadpool_service object.

I.e.:  lifetime(threadpool_service) <= lifetime(threadpool workers) <= lifetime(threadpool impl).

--

Threadpool create/destroy test.
Threadpool self-destruct test.

Self destruct is behaving a bit weird with respect to the program state...
Maybe it's an OS issue, as the suicide code has to detach the worker thread?