Re-implement for intrusively linked iterators.
By linking iterators into the list, the algorithm becomes easier to write.
Each time traversal fails, there is an iterator from which we can restart
the operation.  Iterators are owned by a stack element and linking/unlinking
of the iterator is entirely under control of the thread that owns it.