Fix rare lockup.
When unlink_aid_ cannot change ``b->pred_`` from ``x`` to ``a``,
b is still the correct successor of a.
However it is no longer guaranteed that b doesn't point at x
(it is an indirect successor and may have a pred pointer skipping elements).

In that case, it is important to process b.